requirements.txt
certifi==2018.10.15
cloudinary==1.14.0
Django==2.1
django-cors-headers==2.4.0
django-tinymce==2.7.0
djangorestframework==3.7.7
gunicorn==19.9.0
mock==2.0.0
pbr==5.1.0
Pillow==5.3.0
psycopg2==2.7.3.2
pytz==2018.7
six==1.11.0
urllib3==1.24
django-rest-multiple-models
